refactor(emails): Rename the app, remove the status return and add the function to get store config
