+++ packages/apps/emails/README.md
@@ -1,0 +1,1 @@
+# `@cloudcommerce/app-emails`
