+++ packages/apps/emails/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@cloudcommerce/app-send-transactional-emails",
+  "name": "@cloudcommerce/app-emails",
   "type": "module",
   "version": "0.0.125",
   "description": "E-Com Plus Cloud Commerce default transactional email sending app",
@@ -14,7 +14,7 @@
   "bugs": {
     "url": "https://github.com/ecomplus/cloud-commerce/issues"
   },
-  "homepage": "https://github.com/ecomplus/cloud-commerce/tree/main/packages/apps/send-transactional-emails#readme",
+  "homepage": "https://github.com/ecomplus/cloud-commerce/tree/main/packages/apps/emails#readme",
   "scripts": {
     "build": "sh ../../../scripts/build-lib.sh"
   },
@@ -23,7 +23,6 @@
     "@cloudcommerce/emails": "workspace:*",
     "@cloudcommerce/firebase": "workspace:*",
     "@ecomplus/transactional-mails": "^1.11.1",
-    "axios": "^1.1.3",
     "firebase-admin": "^11.1.0",
     "firebase-functions": "^3.24.1"
   },
