+++ packages/apps/emails/src/functios-lib/trigger-actions.ts
