--- packages/apps/send-transactional-emails/README.md
@@ -1,1 +1,0 @@
-# `@cloudcommerce/app-send-transactional-emails`
